feat: add cluster
